requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
python-dotenv
langchain_community
langchain
numpy
google-generativeai
pydantic
langchain_google_genai